requirements.txt
Pillow
numpy
tensorflow
flask